backend/requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
psycopg2==2.7.6.1
pytz==2018.7
PyYAML==3.13
